chatbot.py
from openai import OpenAI
import os
from dotenv import load_dotenv

load_dotenv()

client = OpenAI(
    api_key = os.getenv("OPEN_API_KEY")
)

def response(messages):
    try:
        response = client.chat.completions.create(
            model="gpt-3.5-turbo",
            messages = messages
        )
        return response.choices[0].message.content
    except Exception as e:
        return f"error {str(e)}"